feat: useSurveyResult 훅을 사용하여 SurveyResult 컴포넌트의 상태 관리 및 API 호출 로직 리팩토링
